Bugfix & simplification
mmu.store32 was storing 8 bytes, oops (by inspection).

mmu.store{32,64} should both take i64 values; the former just ignores
the upper half.  We are moving towards keeping all values as i64 to
minimize the amount of casting (and i64 leads to less casting than
u64).

+++ src/cpu.rs
@@ -2673,7 +2673,7 @@ const INSTRUCTIONS: [Instruction; INSTRUCTION_NUM] = [
         operation: |cpu, word, _address| {
             let f = parse_format_s(word);
             cpu.mmu
-                .store32(cpu.x[f.rs1].wrapping_add(f.imm), cpu.f[f.rs2] as u32)
+                .store32(cpu.x[f.rs1].wrapping_add(f.imm), cpu.f[f.rs2])
         },
         disassemble: dump_format_s,
     },
